partial comment fixing
