--- src/components/molecules/VenueHero/VenueHero.types.ts
@@ -1,10 +1,0 @@
-export interface DetailsPreviewProps {
-  bannerImageUrl?: string;
-  logoImageUrl?: string;
-  name?: string;
-  subtitle?: string;
-  description?: string;
-  large?: boolean;
-  venueId: string;
-  showEdit?: boolean;
-}
