+++ src/components/molecules/VenueHero/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./VenueHero";
+export { VenueHero } from "./VenueHero";
